requirements.txt
numpy
opencv-python
Pillow
requests
streamlit 
